#237; moves sshKey integration to POST /api/workflow.
