+++ api/workflow/initialize.js
@@ -5,6 +5,8 @@ module.exports = self;
 
 var async = require('async');
 var _ = require('underscore');
+var path = require('path');
+var fs = require('fs-extra');
 
 var APIAdapter = require('../../common/APIAdapter.js');
 var envHandler = require('../../common/envHandler.js');
@@ -37,7 +39,11 @@ function initialize(req, res) {
       _getState.bind(null, bag),
       _initializeState.bind(null, bag),
       _getRedis.bind(null, bag),
-      _initializeRedis.bind(null, bag)
+      _initializeRedis.bind(null, bag),
+      _getSSHKeysIntegration.bind(null, bag),
+      _readPublicSSHKey.bind(null, bag),
+      _readPrivateSSHKey.bind(null, bag),
+      _saveSSHKeys.bind(null, bag)
     ],
     function (err) {
       logger.info(bag.who, 'Completed');
@@ -292,6 +298,95 @@ function _initializeRedis(bag, next) {
       if (err)
         return next(
           new ActErr(who, err.id || ActErr.OperationFailed, err)
+        );
+
+      return next();
+    }
+  );
+}
+
+function _getSSHKeysIntegration(bag, next) {
+  var who = bag.who + '|' +  _getSSHKeysIntegration.name;
+  logger.verbose(who, 'Inside');
+
+  var query = 'name=sshKeys&masterName=ssh-key';
+  bag.apiAdapter.getSystemIntegrations(query,
+    function (err, systemIntegrations) {
+      if (err)
+        return next(
+          new ActErr(who, ActErr.OperationFailed,
+            'Failed to get system integrations: ' + util.inspect(err))
+        );
+
+      if (!systemIntegrations.length)
+        bag.saveSSHKeys = true;
+
+      return next();
+    }
+  );
+}
+
+
+function _readPublicSSHKey(bag, next) {
+  if (!bag.saveSSHKeys) return next();
+  var who = bag.who + '|' + _readPublicSSHKey.name;
+  logger.verbose(who, 'Inside');
+
+  var publicKey = path.join(global.config.configDir, 'machinekey.pub');
+  fs.readFile(publicKey,
+    function (err, data) {
+      if (err)
+        return next(
+          new ActErr(who, ActErr.OperationFailed, err)
+        );
+
+      bag.publicSSHKey = data.toString();
+
+      return next();
+    }
+  );
+}
+
+function _readPrivateSSHKey(bag, next) {
+  if (!bag.saveSSHKeys) return next();
+  var who = bag.who + '|' + _readPrivateSSHKey.name;
+  logger.verbose(who, 'Inside');
+
+  var publicKey = path.join(global.config.configDir, 'machinekey');
+  fs.readFile(publicKey,
+    function (err, data) {
+      if (err)
+        return next(
+          new ActErr(who, ActErr.OperationFailed, err)
+        );
+
+      bag.privateSSHKey = data.toString();
+
+      return next();
+    }
+  );
+}
+
+function _saveSSHKeys(bag, next) {
+  if (!bag.saveSSHKeys) return next();
+  var who = bag.who + '|' +  _saveSSHKeys.name;
+  logger.verbose(who, 'Inside');
+
+  var postObject = {
+    name: 'sshKeys',
+    masterName: 'ssh-key',
+    data: {
+      publicKey: bag.publicSSHKey,
+      privateKey: bag.privateSSHKey
+    }
+  };
+
+  bag.apiAdapter.postSystemIntegration(postObject,
+    function (err) {
+      if (err)
+        return next(
+          new ActErr(who, ActErr.OperationFailed,
+            'Failed to create system integration: ' + util.inspect(err))
         );
 
       return next();
